Save Death.kt.

+++ mini/src/main/java/zero/zero/handler/Death.kt
@@ -17,22 +17,32 @@ init { //the constructor of this handler
 
     @EventHandler //says that this is an event handler
     fun onPlayerDead(e: PlayerDeathEvent) {//creating the on player death event function and importing the death class
+    val fileName = "game.txt"// creating the file name 
+    val actualFile = File(fileName)
+    if(actualFile.exist() && actualFile.isFile){
+        val reader = Scanner(actualFile)
+        val data = reader.nextLine()
+        val dataInt = data.toInt()
         val name:String = e.player.name//name variable to store the player name
+        if(dataInt == 0){
         if(p.location.world.name.endsWith("overworld")){
-        e.player.sendMessage("$name died an you all won this match")//send the message of who won the match
+        
+            e.player.sendMessage("$name died an you all won this match")//send the message of who won the match
         //Thread.sleep(10000)//waiting 10sec before tp the player
-        Bukkit.dispatchCommand(e.player, "/server lobby")//sending the player to the lobby
+            Bukkit.dispatchCommand(e.player, "/server lobby")//sending the player to the lobby
             //some code here
 
-        if (e.player.killer is Player) {//checking if the entity killer is player
+            if (e.player.killer is Player) {//checking if the entity killer is player
             //some code here
-            val nameKiller:String = e.player.name//name killer saving the name of the killer to display it in the chat
-            e.player.sendMessage(nameKiller + "kill is team mate")// says who kill who
+                val nameKiller:String = e.player.name//name killer saving the name of the killer to display it in the chat
+                e.player.sendMessage(nameKiller + "kill is team mate")// says who kill who
             //Thread.sleep(10000)//waiting 10 sec before sending it
             //Bukkit.dispatchCommand(e.player, "/server lobby")//sending the player to the lobby
 
         }
 
     }
+}
+}
 }
 }
